refactor(file-storage): cut service/mod.rs HK below 1M by consolidating submodules
The earlier ≤600-SLOC split fragmented `impl FileService` into ~10 tiny
`service/*.rs` files; each `use super::FileService` is a fan_in edge to
`service/mod.rs`, and since HK = sloc × (fan_in × fan_out)² the inflated fan_in
dominated (HK ~2.19M, the #1 hub).

Consolidate the over-fragmentation into the fewest cohesive files that still fit
the 600-SLOC budget (write-path, read-path, backend groups), cutting the number
of `use super::FileService` submodule edges. The path split is preserved (files
stay ≤600 SLOC); only the surplus fragmentation is removed — no change to real
coupling.

service/mod.rs HK ~2,192,832 -> 726,327 (below the 1M target). 168 tests pass,
clippy (-D warnings) + fmt clean.
