+++ gears/file-storage/file-storage/src/domain/service/read_ops.rs
@@ -1,16 +1,152 @@
-//! File and version deletion flows.
+//! Read-only queries (file, metadata, versions) and version-lifecycle operations
+//! (download URL issuance, version listing, restore, and deletion).
 
 use toolkit_security::{AccessScope, SecurityContext};
 use uuid::Uuid;
+
+use file_storage_sdk::{CustomMetadataEntry, File, FileVersion, OwnerFilter};
 
 use crate::domain::audit::AuditOperation;
 use crate::domain::authz::actions;
 use crate::domain::error::DomainError;
 use crate::domain::etag;
-use crate::domain::service::FileService;
+use crate::domain::service::{DownloadTicket, FileService};
 use crate::infra::external_clients::UsageDelta;
 
 impl FileService {
+    // ── reads ─────────────────────────────────────────────────────────────────
+
+    /// Get a file's metadata.
+    pub async fn get_file(
+        &self,
+        ctx: &SecurityContext,
+        file_id: Uuid,
+    ) -> Result<File, DomainError> {
+        let prefetch = Self::tenant_scope(ctx);
+        let file = self.store.require_file(&prefetch, file_id).await?;
+        let scope = self
+            .authorizer
+            .authorize(ctx, actions::READ, &file.gts_file_type, Some(file_id))
+            .await?;
+        self.store.require_file(&scope, file_id).await
+    }
+
+    /// Get a file plus its custom metadata.
+    pub async fn get_file_with_metadata(
+        &self,
+        ctx: &SecurityContext,
+        file_id: Uuid,
+    ) -> Result<(File, Vec<CustomMetadataEntry>), DomainError> {
+        let file = self.get_file(ctx, file_id).await?;
+        let meta = self.store.list_metadata(file_id).await?;
+        Ok((file, meta))
+    }
+
+    /// List files for a mandatory owner filter, offset-paginated.
+    pub async fn list_files(
+        &self,
+        ctx: &SecurityContext,
+        owner: OwnerFilter,
+        limit: Option<u64>,
+        offset: u64,
+    ) -> Result<Vec<File>, DomainError> {
+        // Authorize (access gate), then always tenant-scope the query so the
+        // tenant boundary holds regardless of the PDP's returned constraints.
+        self.authorizer
+            .authorize(ctx, actions::READ, "", None)
+            .await?;
+        let limit = limit
+            .unwrap_or(self.cfg.default_page_size)
+            .min(self.cfg.max_page_size);
+        self.store
+            .list_files(&Self::tenant_scope(ctx), owner, limit, offset)
+            .await
+    }
+
+    // ── pub(crate) accessors for DataPlaneService ─────────────────────────────
+
+    /// Fetch a single version by `(file_id, version_id)` — delegated to the
+    /// data plane so it does not need to hold a direct `Store` reference.
+    pub(crate) async fn get_version(
+        &self,
+        file_id: uuid::Uuid,
+        version_id: uuid::Uuid,
+    ) -> Result<Option<file_storage_sdk::FileVersion>, crate::domain::error::DomainError> {
+        self.store.get_version(file_id, version_id).await
+    }
+
+    // ── download + versioning ─────────────────────────────────────────────────
+
+    /// `GET /files/{id}/download-url`: issue a signed download URL pinned to the
+    /// current content (or a specific `version_id`).
+    pub async fn download_url(
+        &self,
+        ctx: &SecurityContext,
+        file_id: Uuid,
+        version_id: Option<Uuid>,
+    ) -> Result<DownloadTicket, DomainError> {
+        let prefetch = Self::tenant_scope(ctx);
+        let file = self.store.require_file(&prefetch, file_id).await?;
+        let _scope = self
+            .authorizer
+            .authorize(ctx, actions::READ, &file.gts_file_type, Some(file_id))
+            .await?;
+
+        let target = match version_id {
+            Some(v) => v,
+            None => file
+                .content_id
+                .ok_or_else(|| DomainError::conflict("file has no bound content yet"))?,
+        };
+        let version = self
+            .store
+            .get_version(file_id, target)
+            .await?
+            .ok_or_else(|| DomainError::version_not_found(file_id, target))?;
+
+        if version.status != file_storage_sdk::VersionStatus::Available {
+            return Err(DomainError::conflict(
+                "cannot issue a download URL for a version whose upload has not been finalized",
+            ));
+        }
+
+        let download_url =
+            self.build_download_url(file_id, target, version.backend_id, version.backend_path)?;
+        Ok(DownloadTicket {
+            download_url,
+            etag: etag::content_etag(file_id, target),
+            version_id: target,
+        })
+    }
+
+    /// List all versions of a file.
+    pub async fn list_versions(
+        &self,
+        ctx: &SecurityContext,
+        file_id: Uuid,
+    ) -> Result<Vec<FileVersion>, DomainError> {
+        let prefetch = Self::tenant_scope(ctx);
+        let file = self.store.require_file(&prefetch, file_id).await?;
+        let _scope = self
+            .authorizer
+            .authorize(ctx, actions::READ, &file.gts_file_type, Some(file_id))
+            .await?;
+        self.store.list_versions(file_id).await
+    }
+
+    /// Restore a prior version as current (a rebind: pointer swap, no re-upload).
+    pub async fn restore_version(
+        &self,
+        ctx: &SecurityContext,
+        file_id: Uuid,
+        version_id: Uuid,
+    ) -> Result<file_storage_sdk::File, DomainError> {
+        let file = self.get_file(ctx, file_id).await?;
+        let if_match = etag::etag_for(&file);
+        self.bind(ctx, file_id, version_id, if_match.as_deref())
+            .await
+    }
+
     // ── delete ──────────────────────────────────────────────────────────────────
 
     /// `DELETE /files/{id}`: remove the file and all versions (FK cascade) under
