--- gears/file-storage/file-storage/src/domain/service/versioning.rs
@@ -1,85 +1,0 @@
-//! Version listing, download URL issuance, and version restore.
-
-use toolkit_security::SecurityContext;
-use uuid::Uuid;
-
-use file_storage_sdk::FileVersion;
-
-use crate::domain::authz::actions;
-use crate::domain::error::DomainError;
-use crate::domain::etag;
-use crate::domain::service::{DownloadTicket, FileService};
-
-impl FileService {
-    // ── download + versioning ─────────────────────────────────────────────────
-
-    /// `GET /files/{id}/download-url`: issue a signed download URL pinned to the
-    /// current content (or a specific `version_id`).
-    pub async fn download_url(
-        &self,
-        ctx: &SecurityContext,
-        file_id: Uuid,
-        version_id: Option<Uuid>,
-    ) -> Result<DownloadTicket, DomainError> {
-        let prefetch = Self::tenant_scope(ctx);
-        let file = self.store.require_file(&prefetch, file_id).await?;
-        let _scope = self
-            .authorizer
-            .authorize(ctx, actions::READ, &file.gts_file_type, Some(file_id))
-            .await?;
-
-        let target = match version_id {
-            Some(v) => v,
-            None => file
-                .content_id
-                .ok_or_else(|| DomainError::conflict("file has no bound content yet"))?,
-        };
-        let version = self
-            .store
-            .get_version(file_id, target)
-            .await?
-            .ok_or_else(|| DomainError::version_not_found(file_id, target))?;
-
-        if version.status != file_storage_sdk::VersionStatus::Available {
-            return Err(DomainError::conflict(
-                "cannot issue a download URL for a version whose upload has not been finalized",
-            ));
-        }
-
-        let download_url =
-            self.build_download_url(file_id, target, version.backend_id, version.backend_path)?;
-        Ok(DownloadTicket {
-            download_url,
-            etag: etag::content_etag(file_id, target),
-            version_id: target,
-        })
-    }
-
-    /// List all versions of a file.
-    pub async fn list_versions(
-        &self,
-        ctx: &SecurityContext,
-        file_id: Uuid,
-    ) -> Result<Vec<FileVersion>, DomainError> {
-        let prefetch = Self::tenant_scope(ctx);
-        let file = self.store.require_file(&prefetch, file_id).await?;
-        let _scope = self
-            .authorizer
-            .authorize(ctx, actions::READ, &file.gts_file_type, Some(file_id))
-            .await?;
-        self.store.list_versions(file_id).await
-    }
-
-    /// Restore a prior version as current (a rebind: pointer swap, no re-upload).
-    pub async fn restore_version(
-        &self,
-        ctx: &SecurityContext,
-        file_id: Uuid,
-        version_id: Uuid,
-    ) -> Result<file_storage_sdk::File, DomainError> {
-        let file = self.get_file(ctx, file_id).await?;
-        let if_match = etag::etag_for(&file);
-        self.bind(ctx, file_id, version_id, if_match.as_deref())
-            .await
-    }
-}
